Image Upload Ability (FB Storage) etc.
Firstly added a promise, which checks if the user is logged in.
Then coded the logic to upload an image. This firstly uploads to firebase storage, then sets the pictureURL in the database which is read and used to load in the profile picture.
I have my dashboard pulling in the users profile picture too.
I finally worked out how to get the latest workout by timestamp and will use this to update progress.